+++ CHANGELOG.md
@@ -18,3 +18,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Issue #11: Add gyroscope and accelerometer for motion traking
 - Issue #13: Add direct socket connection
 - Issue #16: Motion control through trim
+- Issue #18: Automatic send of status and asset